AppBackend/requirements/development.txt
-r base.txt
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.6.0
black==23.11.0
flake8==6.1.0